feat(types): use `Decimal` as amount and price

+++ exchange-types/src/order_type.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 /// The order type you specify will influence which other order parameters are
 /// required as well as how your order will be executed by the matching engine.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -10,14 +12,14 @@ pub enum OrderType {
     /// to buy or sell, and the price is the price per bitcoin. The limit order
     /// will be filled at the price specified or better.
     Limit {
-        limit_price: u64,
+        limit_price: Decimal,
         /// Time in force policies provide guarantees about the lifetime of an
         /// [order](Order).
         #[cfg_attr(feature = "serde", serde(default))]
         time_in_force: TimeInForce,
-        amount: u64,
+        amount: Decimal,
         #[cfg_attr(feature = "serde", serde(default))]
-        filled: u64,
+        filled: Decimal,
     },
     /// Market orders differ from limit orders in that they provide no pricing
     /// guarantees. They however do provide a way to buy or sell specific
@@ -30,9 +32,9 @@ pub enum OrderType {
         /// is considered a fill or kill order.
         #[cfg_attr(feature = "serde", serde(default))]
         all_or_none: bool,
-        amount: u64,
+        amount: Decimal,
         #[cfg_attr(feature = "serde", serde(default))]
-        filled: u64,
+        filled: Decimal,
     },
 }
